backend/server/pyproject.toml
[tool.poetry]
name = "findplace"
version = "0.1.0"
description = ""
authors = ["YaroslavIonin <ioninyaroslav22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"
psycopg = "^3.1.13"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.1"
django-admin-geomap = "^1.6"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
